Move auth logic into mixin classes
- also ember-osf-models to ember-osf
