+++ app/mixins/osf-login-controller.js
@@ -1,0 +1,1 @@
+export {default} from 'ember-osf/mixins/osf-login-controller';
