+++ app/mixins/osf-login-route.js
@@ -1,0 +1,1 @@
+export {default} from 'ember-osf/mixins/osf-login-route';
