+++ app/models/node.js
@@ -1,1 +1,1 @@
-export { default } from 'ember-osf-models/models/node';
+export { default } from 'ember-osf/models/node';
