+++ app/serializers/application.js
@@ -1,1 +1,1 @@
-export { default } from 'ember-osf-models/serializers/application';
+export { default } from 'ember-osf/serializers/application';
