+++ bower.json
@@ -1,5 +1,5 @@
 {
-  "name": "ember-osf-models",
+  "name": "ember-osf",
   "dependencies": {
     "ember": "2.3.0",
     "ember-cli-shims": "0.1.0",
