+++ tests/dummy/app/router.js
@@ -2,12 +2,13 @@ import Ember from 'ember';
 import config from './config/environment';
 
 const Router = Ember.Router.extend({
-  location: config.locationType
+    location: config.locationType
 });
 
 Router.map(function() {
-  this.route('nodes');
-  this.route('login');
+    this.route('index', {path: '/'});
+    this.route('nodes');
+    this.route('login');
 });
 
 export default Router;
