+++ tests/dummy/app/routes/index.js
@@ -1,5 +1,4 @@
 import Ember from 'ember';
 import AuthenticatedRouteMixin from 'ember-simple-auth/mixins/authenticated-route-mixin';
 
-export default Ember.Route.extend(AuthenticatedRouteMixin, {
-});
+export default Ember.Route.extend(AuthenticatedRouteMixin, {});
