+++ tests/dummy/app/routes/login.js
@@ -1,27 +1,4 @@
 import Ember from 'ember';
-import UnauthenticatedRouteMixin from 'ember-simple-auth/mixins/unauthenticated-route-mixin';
+import OsfLoginRouteMixin from 'ember-osf/mixins/osf-login-route';
 
-import config from 'ember-get-config';
-
-export default Ember.Route.extend(UnauthenticatedRouteMixin, {
-    session: Ember.inject.service(),
-    beforeModel() {
-        var accessToken;
-        if (config.DEV) {
-            accessToken = config.OSF.accessToken;
-        }
-        else {
-            // Acquire an OSF access token, then exchange it for a Jam token
-            var hash = window.location.hash.substring(1).split('&').map(function(str) {
-                return this[str.split('=')[0]] = str.split('=')[1], this;
-            }.bind({}))[0];
-            if (!hash || !hash.access_token) {
-                return null;
-            }
-            window.location.hash = '';
-            accessToken = hash.access_token;
-        }
-        return this.get('session').authenticate('authenticator:osf-token', accessToken)
-            .then(() => this.transitionTo('index'));
-    }
-});
+export default Ember.Route.extend(OsfLoginRouteMixin);
